remove redux routing, move form to separate container,adjust route calls
