--- src/actions/routing.ts
@@ -1,8 +1,0 @@
-export const ROUTE_TO_PAGE = 'ROUTE_TO_PAGE';
-
-export function routeToPage(page: string) {
-  return {
-    page,
-    type: ROUTE_TO_PAGE
-  };
-}
