+++ src/components/Diagnosis.tsx
