+++ src/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { HashRouter as Router } from 'react-router-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 import configureStore from './store/configureStore';
 
 import App from './containers/App';
@@ -13,11 +13,11 @@ import registerServiceWorker from './registerServiceWorker';
 const store = configureStore();
 
 ReactDOM.render(
-  <Provider store={store}>
-    <Router>
+  <Router>
+    <Provider store={store}>
       <App />
-    </Router>
-  </Provider>,
+    </Provider>
+  </Router>,
   document.getElementById('root') as HTMLElement
 );
 registerServiceWorker();
